+++ src/ExpressiveInstaller/Resources/config/routes-aura-router.php
@@ -1,13 +1,17 @@
 <?php
 
+use App\Action;
+use Zend\Expressive\Router\AuraRouter;
+use Zend\Expressive\Router\RouterInterface;
+
 return [
     'dependencies' => [
         'invokables' => [
-            Zend\Expressive\Router\RouterInterface::class => Zend\Expressive\Router\AuraRouter::class,
-            App\Action\PingAction::class => App\Action\PingAction::class,
+            RouterInterface::class => AuraRouter::class,
+            Action\PingAction::class => Action\PingAction::class,
         ],
         'factories' => [
-            App\Action\HomePageAction::class => App\Action\HomePageFactory::class,
+            Action\HomePageAction::class => Action\HomePageFactory::class,
         ],
     ],
 ];